Measurement data type changed to unsigned int

+++ Server/Mqtt/MeasurementsMqttClient.cs
@@ -51,7 +51,7 @@ namespace RemoteCarDiagz.Server.Mqtt
             var metricName = arg.ApplicationMessage.Topic.Substring(arg.ApplicationMessage.Topic.LastIndexOf('/') + 1);
             var metric = Metrics.CreateGauge(metricName, arg.ApplicationMessage.Topic);
             using MemoryStream stream = new(arg.ApplicationMessage.Payload);
-            var json = await JsonSerializer.DeserializeAsync<byte>(stream);
+            var json = await JsonSerializer.DeserializeAsync<UInt16>(stream);
             metric.Set(json);
             _logger.LogInformation("Message received: {0}, value: {1}, with metric {2}", arg.ApplicationMessage.Topic, json, metricName);
         }
